feat(runtime-c-api) Extract the `module` module.
